Promote workspace mod to directory
